rm bugs
