+++ scdiffeq/__init__.py
@@ -1,13 +1,11 @@
 
 __module_name__ = "__init__.py"
-__version__ = __Version__ = "0.0.47rc0"
+__version__ = __Version__ = "0.0.47rc1"
 __doc__ = """Top-level __init__ for the scdiffeq package."""
-__author__ = ", ".join(["Michael E. Vinyard", "Anders Rasmussen", "Ruitong Li"])
+__author__ = ", ".join(["Michael E. Vinyard"])
 __email__ = ", ".join(
     [
         "mvinyard@broadinstitute.org",
-        "arasmuss@broadinstitute.org",
-        "ruitong@broadinstitute.org",
     ]
 )
