+++ setup.py
@@ -5,7 +5,7 @@ import sys
 
 setuptools.setup(
     name="scdiffeq",
-    version="0.0.47rc0",
+    version="0.0.47rc1",
     python_requires=">3.9.0",
     author="Michael E. Vinyard - Harvard University - Massachussetts General Hospital - Broad Institute of MIT and Harvard",
     author_email="mvinyard@broadinstitute.org",
@@ -32,6 +32,7 @@ setuptools.setup(
         "vinplots>=0.0.75",
         "annoyance==0.0.18",
         "ABCParse==0.0.3",
+        "scdiffeq-plots==0.0.1rc2",
     ],
     classifiers=[
         "Development Status :: 2 - Pre-Alpha",
